+++ src/main/java/sopt/hackerthon/hackerthon/entity/Chat.java
@@ -1,10 +1,8 @@
-package sopt.hackerthon.hackerthon.common.domain;
+package sopt.hackerthon.hackerthon.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @Entity
 @Getter
